Move ClaspSettings, ProjectSettings, and DOTFILE to utils
